#169 Create domain-focused type modules under src/world/types/
