+++ src/world/types/conversation.ts
@@ -1,0 +1,6 @@
+export interface ConversationMessage {
+  role: 'player' | 'assistant';
+  text: string;
+}
+
+export type ActorConversationHistoryByActorId = Record<string, ConversationMessage[]>;
