+++ src/world/types/environment.ts
@@ -1,0 +1,8 @@
+import type { GridPosition } from './grid.js';
+
+export interface Environment {
+  id: string;
+  displayName: string;
+  position: GridPosition;
+  isBlocking: boolean;
+}
